+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 1. (BREAKING) Deprecated legacy type definitions.
 2. (BREAKING) sentenceCase no longer uses a default keyword map.
+3. (BREAKING) Removed legacy value from OmniSearch utils.
 
 ## 1.1.0 - 2019-##-